+++ scripts/ci/core-dialyzer.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+# shellcheck disable=SC1091
+source "$script_dir/_common.sh"
+
+mix dialyzer --format github